Enum cfb::Version

pub enum Version {
    V3,
    V4,
}

The CFB format version to use.

Variants
V3

Version 3, which uses 512-byte sectors.

V4

Version 4, which uses 4096-byte sectors.

Methods

impl Version

fn from_number(number: u16) -> Option<Version>

Returns the version enum for the given version number, or None.

fn number(self) -> u16

Returns the version number for this version.

fn sector_shift(self) -> u16

Returns the sector shift used in this version.

fn sector_len(self) -> usize

Returns the length of sectors used in this version.

use cfb::Version;
assert_eq!(Version::V3.sector_len(), 512);
assert_eq!(Version::V4.sector_len(), 4096);

fn stream_len_mask(self) -> u64

Returns the bitmask used for reading stream lengths in this version.

fn dir_entries_per_sector(self) -> usize

Returns the number of directory entries per sector in this version.