Struct cfb::Entry

pub struct Entry { /* fields omitted */ }

Metadata about a single object (storage or stream) in a compound file.

Methods

impl Entry

fn name(&self) -> &str

Returns the name of the object that this entry represents.

fn path(&self) -> &Path

Returns the full path to the object that this entry represents.

fn is_stream(&self) -> bool

Returns whether this entry is for a stream object (i.e. a "file" within the compound file).

fn is_storage(&self) -> bool

Returns whether this entry is for a storage object (i.e. a "directory" within the compound file), either the root or a nested storage.

fn is_root(&self) -> bool

Returns whether this entry is specifically for the root storage object of the compound file.

fn len(&self) -> u64

Returns the size, in bytes, of the stream that this metadata is for.

fn clsid(&self) -> &[u8; 16]

Returns the CLSID (that is, the object class GUID) for this object. This will always be all zeros for stream objects.

fn state_bits(&self) -> u32

Returns the user-defined bitflags set for this object.

fn created(&self) -> SystemTime

Returns the time when the object that this entry represents was created.

fn modified(&self) -> SystemTime

Returns the time when the object that this entry represents was last modified.